+++ NEWS.md
@@ -1,12 +1,12 @@
 # teal.modules.clinical 0.9.1.9013
 
-* Removed `Show Warnings` modals from modules.
-
 ### Enhancements
 * Added `teal.logger` functionality for logging changes in shiny inputs in all modules.
 * Introduced `ylim` parameter for `tm_g_km` module that controls width of y-axis.
+* Added functionality to `tm_t_events_patyear` to split columns by multiple (nested) variables via the `arm_var` argument.
 
 ### Miscellaneous
+* Removed `Show Warnings` modals from modules.
 * Clarified the documentation specifying whether multiple values can be selected in the `arm_var` argument for each module.
 
 # teal.modules.clinical 0.9.1